requirements.txt
absl-py==2.0.0
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
astunparse==1.6.3
attrs==23.1.0
bidict==0.22.1
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
distro==1.8.0
docopt==0.6.2
Flask==2.3.3
flatbuffers==23.5.26
frozenlist==1.4.1
gast==0.5.4
google-ai-generativelanguage==0.4.0
google-api-core==2.18.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-generativeai==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
h5py==3.10.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
iso8601==2.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
keras==2.15.0
libclang==16.0.6
llvmlite==0.41.1
Markdown==3.5.1
MarkupSafe==2.1.3
ml-dtypes==0.2.0
msgpack==1.0.7
multidict==6.0.4
numba==0.58.1
numpy==1.26.2
oauthlib==3.2.2
openai==1.6.0
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
passlib==1.7.4
pipreqs==0.4.13
platformdirs==4.1.0
pooch==1.8.0
proto-plus==1.23.0
protobuf==4.23.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.14
pycparser==2.21
pycryptodomex==3.19.0
pydantic==2.5.2
pydantic_core==2.14.5
pyseto==1.7.7
python-engineio==4.8.0
python-socketio==5.10.0
replit==3.5.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.4
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
sounddevice==0.4.6
soundfile==0.12.1
soxr==0.3.7
SpeechRecognition==3.10.1
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.2.0
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==1.26.18
wavio==0.0.8
Werkzeug==3.0.1
wrapt==1.14.1
wsproto==1.2.0
yarg==0.1.9
yarl==1.9.4
